# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CPSCOM
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/v1/group/group-list:
    post:
      summary: 'Group List'
      operationId: groupList
      description: '@bodyParam uid string required The id of the user.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  statusCode: 200
                  data:
                    '0':
                      admin: 'Tony nelson'
                      id: AJolDjZVXfywCFtjgTVEBSxWaYPudIUHVMhK
                      name: 'Nurses at Bristol'
                      profile_picture: group/1qQtvEY71EM76wDJsoNXXCK6tWfGWJtlJaPspRaD.png
                    '2':
                      admin: CRClinic
                      id: fRQDiLq0sMbAmc9QMFmUkL07cAWhIaiW3TVX
                      name: 'New Group'
                      profile_picture: group/wiFx4czB5JyNlzTm0nbW5Lhzf0dw2GaRsPe8ZQQZ.png
                properties:
                  status:
                    type: boolean
                    example: true
                  statusCode:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      0:
                        type: object
                        properties:
                          admin:
                            type: string
                            example: 'Tony nelson'
                          id:
                            type: string
                            example: AJolDjZVXfywCFtjgTVEBSxWaYPudIUHVMhK
                          name:
                            type: string
                            example: 'Nurses at Bristol'
                          profile_picture:
                            type: string
                            example: group/1qQtvEY71EM76wDJsoNXXCK6tWfGWJtlJaPspRaD.png
                      2:
                        type: object
                        properties:
                          admin:
                            type: string
                            example: CRClinic
                          id:
                            type: string
                            example: fRQDiLq0sMbAmc9QMFmUkL07cAWhIaiW3TVX
                          name:
                            type: string
                            example: 'New Group'
                          profile_picture:
                            type: string
                            example: group/wiFx4czB5JyNlzTm0nbW5Lhzf0dw2GaRsPe8ZQQZ.png
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: false
                  statusCode: 401
                  error:
                    message:
                      - 'The uid field is required.'
                properties:
                  status:
                    type: boolean
                    example: false
                  statusCode:
                    type: integer
                    example: 401
                  error:
                    type: object
                    properties:
                      message:
                        type: array
                        example:
                          - 'The uid field is required.'
                        items:
                          type: string
      tags:
        - 'Group APIs'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uid:
                  type: string
                  description: ''
                  example: ut
              required:
                - uid
      security: []
...
  /api/v1/group/media-image-download:
    post:
      summary: 'Media Image Download'
      operationId: mediaImageDownload
      description: '@bodyParam url string required The url of the image.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n \"status\": true,\n  \"statusCode\": 200,\n\n}"
        401:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"status\": false,\n\"statusCode\": 401,\n\"error\": {\n\"message\": [\n\"The url field is required.\"\n]\n}"
      tags:
        - 'Group APIs'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: ''
                  example: non
              required:
                - url
      security: []
...